2018-02-26  Dominique d'Humieres <dominiq@gcc.gnu.org>
	PR fortran/32957
	* gfortran.dg/c_f_pointer_shape_tests_2.f03: Use explicit KIND c_int.
	* gfortran.dg/c_f_pointer_shape_tests_4.f03: Likewise.
	* gfortran.dg/c_funloc_tests_3.f03: Likewise.
	* gfortran.dg/c_loc_test.f90: Likewise.
	* gfortran.dg/c_loc_tests_2.f03: Likewise.
	* gfortran.dg/proc_decl_17.f90: Likewise.
	* gfortran.dg/proc_ptr_8.f90: Likewise.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@257985 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gfortran.dg/c_f_pointer_shape_tests_2.f03
@@ -12,7 +12,7 @@ contains
     use, intrinsic :: iso_c_binding
     type(c_ptr), value :: cPtr
     integer(c_int), value :: num_elems
-    integer, dimension(:), pointer :: myArrayPtr
+    integer(c_int), dimension(:), pointer :: myArrayPtr
     integer(c_long_long), dimension(1) :: shape
     integer :: i
     
@@ -28,7 +28,7 @@ contains
     type(c_ptr), value :: cPtr
     integer(c_int), value :: num_rows
     integer(c_int), value :: num_cols
-    integer, dimension(:,:), pointer :: myArrayPtr
+    integer(c_int), dimension(:,:), pointer :: myArrayPtr
     integer(c_long_long), dimension(2) :: shape
     integer :: i,j
     
@@ -46,7 +46,7 @@ contains
     use, intrinsic :: iso_c_binding
     type(c_ptr), value :: cPtr
     integer(c_int), value :: num_elems
-    integer, dimension(:), pointer :: myArrayPtr
+    integer(c_int), dimension(:), pointer :: myArrayPtr
     integer(c_long), dimension(1) :: shape
     integer :: i
     
@@ -61,7 +61,7 @@ contains
     use, intrinsic :: iso_c_binding
     type(c_ptr), value :: cPtr
     integer(c_int), value :: num_elems
-    integer, dimension(:), pointer :: myArrayPtr
+    integer(c_int), dimension(:), pointer :: myArrayPtr
     integer(c_int), dimension(1) :: shape
     integer :: i
     
@@ -76,7 +76,7 @@ contains
     use, intrinsic :: iso_c_binding
     type(c_ptr), value :: cPtr
     integer(c_int), value :: num_elems
-    integer, dimension(:), pointer :: myArrayPtr
+    integer(c_int), dimension(:), pointer :: myArrayPtr
     integer(c_short), dimension(1) :: shape
     integer :: i
     
@@ -91,7 +91,7 @@ contains
     use, intrinsic :: iso_c_binding
     type(c_ptr), value :: cPtr
     integer(c_int), value :: num_elems
-    integer, dimension(:), pointer :: myArrayPtr
+    integer(c_int), dimension(:), pointer :: myArrayPtr
     integer(c_int), dimension(1) :: shape1
     integer(c_long_long), dimension(1) :: shape2
     integer :: i
